chore(example): rename datasource
